Remove obaque and fix extrinsics root calculation
