+++ src/service.rs
@@ -6,7 +6,7 @@ use std::sync::Arc;
 use primitives::ed25519;
 use runtime_primitives::BasicInherentData;
 use transaction_pool::{self, txpool::{Pool as TransactionPool}};
-use shasper_primitives::{ValidatorId, opaque::Block};
+use shasper_primitives::{ValidatorId, Block};
 use shasper_runtime::{self, GenesisConfig, RuntimeApi};
 use substrate_service::{
 	FactoryFullConfiguration, LightComponents, FullComponents, FullBackend,
